move "Enums" code to its own CPP file
