+++ six/modules/c++/six/six.vcxproj.filters
@@ -260,5 +260,8 @@
     <ClCompile Include="source\XmlLite.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="source\Enums.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
 </Project>
